backend/requirement.txt
python-dotenv==1.0.1
pymongo[srv]      
Flask
openai